+++ app/stores/AdaRedemptionStore.js
@@ -32,6 +32,7 @@ export default class AdaRedemptionStore extends Store {
   @observable redeemAdaRequest: Request<Wallet> = new Request(this.api.redeemAda);
   // eslint-disable-next-line
   @observable redeemPaperVendedAdaRequest: Request<RedeemPaperVendedAdaResponse> = new Request(this.api.redeemPaperVendedAda);
+  @observable isRedemptionDisclaimerAccepted = false;
 
   setup() {
     const actions = this.actions.adaRedemption;
@@ -47,6 +48,7 @@ export default class AdaRedemptionStore extends Store {
     actions.adaSuccessfullyRedeemed.listen(this._onAdaSuccessfullyRedeemed);
     actions.closeAdaRedemptionSuccessOverlay.listen(this._onCloseAdaRedemptionSuccessOverlay);
     actions.removeCertificate.listen(this._onRemoveCertificate);
+    actions.acceptRedemptionDisclaimer.listen(this._onAcceptRedemptionDisclaimer);
     ipcRenderer.on(PARSE_REDEMPTION_CODE.SUCCESS, this._onCodeParsed);
     ipcRenderer.on(PARSE_REDEMPTION_CODE.ERROR, this._onParseError);
   }
@@ -72,6 +74,10 @@ export default class AdaRedemptionStore extends Store {
       this.redemptionType = params.redemptionType;
     }
   };
+
+  _onAcceptRedemptionDisclaimer = action(() => {
+    this.isRedemptionDisclaimerAccepted = true;
+  });
 
   _setCertificate = action(({ certificate }) => {
     this.certificate = certificate;